refactor: centralize useToast hook
