+++ src/ui/headless/admin/audit-logs/AdminAuditLogs.tsx
@@ -1,5 +1,5 @@
 import { ReactNode, useEffect, useState } from 'react';
-import { useToast } from '@/ui/primitives/use-toast';
+import { useToast } from '@/lib/hooks/use-toast';
 
 /**
  * Headless AdminAuditLogs component that provides audit log functionality through render props
