+++ src/ui/styled/company/DomainManagement.tsx
@@ -7,7 +7,7 @@ import { Card, CardContent, CardHeader, CardTitle, CardDescription } from '@/ui/
 import { Alert, AlertDescription, AlertTitle } from '@/ui/primitives/alert';
 import { Badge } from '@/ui/primitives/badge';
 import { api } from '@/lib/api/axios';
-import { useToast } from '@/ui/primitives/use-toast';
+import { useToast } from '@/lib/hooks/use-toast';
 import { CheckCircle, Clock, AlertCircle, Loader2, Plus, Trash, Globe, Shield, Star } from 'lucide-react';
 import { Input } from '@/ui/primitives/input';
 import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from '@/ui/primitives/table';
