+++ src/ui/styled/company/NotificationPreferences.tsx
@@ -11,7 +11,7 @@ import { Label } from '@/ui/primitives/label';
 import { Badge } from '@/ui/primitives/badge';
 import { Loader2, Bell, Mail, Plus, Trash, AlertCircle, Info } from 'lucide-react';
 import { api } from '@/lib/api/axios';
-import { useToast } from '@/ui/primitives/use-toast';
+import { useToast } from '@/lib/hooks/use-toast';
 import {
   CompanyNotificationPreference,
   NotificationType,
