+++ src/ui/styled/profile/__tests__/ProfileForm.test.tsx
@@ -41,7 +41,7 @@ vi.mock('@/hooks/auth/useAuth', () => ({
 }));
 
 // Mock the toast component
-vi.mock('@/ui/primitives/use-toast', () => ({
+vi.mock('@/lib/hooks/use-toast', () => ({
   useToast: () => ({ toast: vi.fn() }),
 }));
 vi.mock('@/components/ui/textarea', () => ({
